Modified the way to access the database.
