+++ src/main/java/controller/FrontPageController.java
@@ -19,9 +19,4 @@ public class FrontPageController {
         FrontPageRequestModel frontPageRequestModel = new FrontPageRequestModel();
         this.frontPageInputBoundary.create(frontPageRequestModel);
     }
-
-    public void create(String target){
-        FrontPageRequestModel frontPageRequestModel = new FrontPageRequestModel(target);
-        this.frontPageInputBoundary.create(frontPageRequestModel);
-    }
 }
