+++ src/main/java/use_case_frontpage/FrontPageRequestModel.java
@@ -4,19 +4,8 @@ public class FrontPageRequestModel {
     /**
      * The request model generated by the controller to be passed to the interactor
      * to be processed.
-     * Contain the target who the current user is chatting with.
+     * Request for the presenter to update the chat screen.
      */
-    private final String target;
-
     public FrontPageRequestModel(){
-        this.target = null;
-    }
-
-    public FrontPageRequestModel(String target){
-        this.target = target;
-    }
-
-    public String getTarget() {
-        return target;
     }
 }
